- renamed fiel MovingDirection.ts to moving-direction.enum.ts
