+++ src/components/components/wizard-step.component.spec.ts
@@ -3,7 +3,7 @@ import {async, ComponentFixture, TestBed} from '@angular/core/testing';
 import {WizardStepComponent} from './wizard-step.component';
 import {ViewChild, Component} from '@angular/core';
 import {WizardComponent} from './wizard.component';
-import {MovingDirection} from '../util/MovingDirection';
+import {MovingDirection} from '../util/moving-direction.enum';
 import {WizardNavigationBarComponent} from './wizard-navigation-bar.component';
 import {GoToStepDirective} from '../directives/go-to-step.directive';
 import {By} from '@angular/platform-browser';
