+++ src/components/components/wizard-step.component.ts
@@ -1,5 +1,5 @@
 import {Component, ContentChild, EventEmitter, forwardRef, HostBinding, Input, Output} from '@angular/core';
-import {MovingDirection} from '../util/MovingDirection';
+import {MovingDirection} from '../util/moving-direction.enum';
 import {WizardStep} from '../util/WizardStep';
 import {WizardStepTitleDirective} from '../directives/wizard-step-title.directive';
