+++ src/components/util/index.ts
@@ -2,6 +2,6 @@
  * Created by marc on 04.01.17.
  */
 
-export {MovingDirection} from './MovingDirection';
+export {MovingDirection} from './moving-direction.enum';
 export {WizardStep} from './WizardStep'
 export * from './StepOffset';
